Wait for services to actually stop before exiting.
Remove duplication between BeaconNode and ValidatorNode.
